Add files via upload
🌲🔥 Forest Fire Detection using Deep Learning
This project focuses on building an intelligent and efficient system to detect forest fires from images using deep learning techniques. The main goal is to automate the process of identifying fire in forest regions by analyzing image data and enabling early warning systems that can assist in reducing the devastating impact of wildfires on the environment, wildlife, and human life.